src/vvs_local/requirements.txt
torch==2.7.0
transformers==4.51.3
pandas==2.2.0
pika==1.3.2
rdkit==2024.09.6
tqdm
rich==13.7.1
pyyaml==6.0.1
pyarrow==19.0.1
python-dotenv==1.0.1
duckdb==1.2.2
datasets==2.19.1
usearch==2.17.7